Move ContentModel to its own source file
